Make CMake build system more platform-independent
I was having a lot of problems building the code, because CMake couldn't
find some of the libraries on my system. I've taken the liberty of
modifying the CMake code a little bit to make sure it works on my
machine, and hopefully this will make it more easy to compile on other
people's machines as well!